+++ docs/.vitepress/config.js
@@ -32,6 +32,10 @@ const config = {
           { text: 'Getting Help', link: '/guides/getting-help' },
           { text: 'Steam publishing', link: '/guides/steam-publishing' },
           { text: 'Updating narrat', link: '/guides/updating-narrat' },
+          {
+            text: 'Custom Start Menu Buttons',
+            link: '/guides/custom-start-buttons',
+          },
         ],
       },
       {